schema and operation with DB added
