+++ .env.example
@@ -1,0 +1,1 @@
+DB_HOST
