--- models/contacts.js
@@ -1,56 +1,0 @@
-const fs = require("fs/promises");
-const path = require("path");
-const { nanoid } = require("nanoid");
-
-const contactsPath = path.join(__dirname, "contacts.json");
-
-async function listContacts() {
-  //чтение содержимого файла контактов
-  const data = await fs.readFile(contactsPath, "utf-8");
-  return JSON.parse(data);
-}
-
-async function getContactById(contactId) {
-  // найти контакт по Id
-  const contacts = await listContacts();
-  return contacts.find((item) => item.id === contactId);
-}
-
-async function removeContact(contactId) {
-  // удалить контакт по Id
-  const contacts = await listContacts();
-  const index = contacts.findIndex((item) => item.id === contactId);
-  if (index === -1) return "this contact is not listed";
-  const [result] = contacts.splice(index, 1);
-  await fs.writeFile(contactsPath, JSON.stringify(contacts, null, 2));
-  return result;
-}
-
-async function addContact(data) {
-  // добавить новый контакт в список контактов
-  const newContact = {
-    id: nanoid(),
-    ...data,
-  };
-  const contacts = await listContacts();
-  contacts.push(newContact);
-  await fs.writeFile(contactsPath, JSON.stringify(contacts, null, 2));
-  return newContact;
-}
-
-async function updateContact(contactId, data) {
-  const contacts = await listContacts();
-  const index = contacts.findIndex((contact) => contact.id === contactId);
-  if (index === -1) return null;
-  contacts[index] = { id: contactId, ...data };
-  await fs.writeFile(contactsPath, JSON.stringify(contacts, null, 2));
-  return contacts[index];
-}
-
-module.exports = {
-  listContacts,
-  getContactById,
-  removeContact,
-  addContact,
-  updateContact,
-};
