+++ models/schemas/contact.js
@@ -1,0 +1,28 @@
+const { Schema, model } = require("mongoose");
+
+const contactSchema = new Schema(
+  {
+    name: {
+      type: String,
+      required: [true, "Set name for contact"],
+    },
+    email: {
+      type: String,
+      require: true,
+      match: /.+@.+\..+/i,
+    },
+    phone: {
+      type: String,
+      required: true,
+    },
+    favorite: {
+      type: Boolean,
+      default: false,
+    },
+  },
+  { versionKey: false, timestamps: true }
+);
+
+const Contact = model("contact", contactSchema);
+
+module.exports = Contact;
